backend/flask-app/requirements.txt
eventlet==0.39.0
Flask==3.1.0
python-dotenv==1.0.1
fixtures
Flask-Cors==5.0.0
maturin>=0.13,<0.15
eclipse-zenoh== 0.11.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.9.0.post0
psycopg2-binary
celery[sqlalchemy]==5.4.0
wrapt==1.14.0
pandas==2.2.3
ydata-profiling==4.12.2
great-expectations==1.3.10
openpyxl
lxml==5.2.2
pyarrow==15.0.2 
geopandas==0.13.2
langchain==0.3.22
langchain-community==0.0.30
langchain-ollama